# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ODEL HUB TON Pay API
  version: 0.1.0
  description: |
    HTTP API for the ODEL HUB tuition payment platform (UGX fees, TON settlement, MongoDB persistence).
    Deployed as **Next.js Route Handlers** on **Vercel**. Admin routes require JWT cookie (`odelhub_admin`) from `POST /api/auth/login`.

    Optional ops env: **RESEND_API_KEY**, **RESEND_FROM** (receipt email); **BRIDGE_WEBHOOK_URL**, **BRIDGE_WEBHOOK_SECRET**, **EXCHANGE_SWAP_URL** (UGX→TON bridge after MoMo confirm).
servers:
  - url: https://{vercel-deployment}
    variables:
      vercel-deployment:
        default: your-app.vercel.app
...
paths:
...
  /api/programmes/{code}/quote:
    get:
      tags: [Programmes]
      summary: Fee quote for year/semester + TON estimate
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
        - name: year
          in: query
          required: true
          schema: { type: integer, minimum: 1, maximum: 6 }
        - name: semester
          in: query
          required: true
          schema: { type: integer, minimum: 1, maximum: 2 }
      responses:
        "200":
          description: Quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteResponse"
        "400": { description: Missing query }
        "404": { description: Programme or fee line not found }
...
  /api/students/{id}:
    get:
      tags: [Students]
      summary: Student detail + payments (admin)
      security:
        - AdminCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
        "401": { description: Unauthorized }
        "404": { description: Not found }
...
components:
  securitySchemes:
    AdminCookie:
      type: apiKey
      in: cookie
      name: odelhub_admin
    PartnerApiKey:
      type: http
      scheme: bearer
      bearerFormat: odelhub_live
      description: Partner API key from Master Admin (scope payments:create required)
  schemas:
...
    QuoteResponse:
      type: object
      properties:
        programmeCode: { type: string }
        programmeName: { type: string }
        year: { type: integer }
        semester: { type: integer }
        tuitionUgx: { type: number }
        functionalFeesUgx: { type: number }
        totalUgx: { type: number }
        ugxPerTon: { type: number }
        rateSource: { type: string }
        tonAmount: { type: number }
        destinationWallet: { type: string }
...